Promote CIndicesTag/DispatchRegime to CTurbSolver, shared by SA and (soon) SST
Both models need the identical regime/NEMO -> FlowIndices dispatch; keeping
it on CTurbSASolver alone would mean either duplicating it on
CTurbSSTSolver or reaching across sibling classes. Header-defined (not just
declared) on the shared CTurbSolver base, since it is a template over a
deduced, unnameable lambda type and needs to be usable from more than one
translation unit (CTurbSASolver.cpp and, next, CTurbSSTSolver.cpp).

Pure refactor: bit-identical to the pre-refactor binary on turb_ONERAM6.
